// src/lib/types.d.ts
export type Path = Array<string | number>

/**
 * @typedef {'after' | 'key' | 'value' | 'append'} CaretType
 */
export type CaretType = 'after' | 'key' | 'value' | 'append'

/**
 * @typedef {{
 *   start: number,
 *   end: number
 * }} VisibleSection
 */
export type VisibleSection = Record<'start' | 'end', number>

/**
 * @typedef {{
 *   path: Path,
 *   type: CaretType
 * }} CaretPosition
 */
export interface CaretPosition {
  path: Path
  type: CaretType
}
// ...
export interface MultiSelection {
  type: 'multi'
  paths: Path[]
  anchorPath: Path
  focusPath: Path
  pathsMap: Record<string, string | boolean>
}

/**
 * @typedef {{type: 'after', path: Path, anchorPath: Path, focusPath: Path}} AfterSelection
 */
export interface AfterSelection {
  type: 'after'
  path: Path
  anchorPath: Path
  focusPath: Path
}

/**
 * @typedef {{type: 'inside', path: Path, anchorPath: Path, focusPath: Path}} InsideSelection
 */
export interface InsideSelection {
  type: 'inside'
  path: Path
  anchorPath: Path
  focusPath: Path
}

/**
 * @typedef {{type: 'key', path: Path, anchorPath: Path, focusPath: Path, edit?: boolean}} KeySelection
 */
export interface KeySelection {
  type: 'key'
  path: Path
  anchorPath: Path
  focusPath: Path
  edit?: boolean
}

/**
 * @typedef {{type: 'value', path: Path, anchorPath: Path, focusPath: Path, edit?: boolean}} ValueSelection
 */
export interface ValueSelection {
  type: 'value'
  path: Path
  anchorPath: Path
  focusPath: Path
  edit?: boolean
}

/**
 * @typedef {MultiSelection | AfterSelection | InsideSelection | KeySelection | ValueSelection} Selection
 */
export type Selection =
  | MultiSelection
  | AfterSelection
  | InsideSelection
  | KeySelection
  | ValueSelection